Add option to allow skipping analysis if no face is found
It removes enforce_detection from DeepFace analyze and replaces it with
on_missing_face. The old behaviour is retained as 'enforce' results in
the same behaviour as enforce_detection=True and 'error' results in the
same behaviour as enforce_detection=False. The third option, 'skip',
allows to just return an empty dataset.

This is especially useful when analysing multiple images where analysis
should not be stopped if any one contains no face, but it is also
desirable to know which of the images didn't contain a face (without
relying on bounds checking of the returned region).